telegram_bot/requirements.txt
python-telegram-bot==13.15
requests==2.32.3
python-dotenv==1.0.1
solana==0.31.0